// src/app/models/paiement_test.ts
import { Facteur } from "./facteur";
import { Paiement } from "./paiement";

export class Paiement_test{
    id:number;
    idStructure: number;
    libelleStructure: String;
    mois: number;
    annee: number;
    datepaie: Date;
    netapayer: number;
    totaltva: number;
    totalMontantHt: number;
    totalRemise: number;
    totalTtc: number;
    paiement: any; 
    
    
}